Refactor service imports and add autoflake and coverage modules
- Update import paths for process_coverage_json and process_autoflake_results to reflect new module structure.
- Introduce new autoflake.py and coverage.py modules for processing autoflake results and coverage data, respectively.
- Enhance logging setup in the new modules for better traceability and error handling.
- Update test files to align with the new module imports, ensuring all tests remain functional.
