+++ src/mcp_suite/servers/qa/service/__init__.py
@@ -5,7 +5,7 @@
 # Import the centralized logger
 from mcp_suite.servers.qa import logger
 
-from .coverage_service import process_coverage_json
+from .coverage import process_coverage_json
 from .pytest_service import process_pytest_results
 
 __all__ = ["process_coverage_json", "process_pytest_results", "logger"]
